requirements.txt
dockerspawner
jupyterhub-dummyauthenticator
oauthenticator
